fix shader ivec

+++ GLSLRayTracer/GLSLRayTracer/Chapter5/PathTracePS.glsl
@@ -2,7 +2,7 @@
 in vec2 screenCoord;
 
 uniform sampler2D envMap;
-uniform ivec2 screenSize;
+uniform vec2 screenSize;
 
 out vec4 FragColor;
